Reorganized src/lib/db/resume files again to mirror api structure.
