+++ src/app/api/mcp/resume/route.ts
@@ -1,8 +1,8 @@
 // Dependencies.
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js"
 import { WebStandardStreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/webStandardStreamableHttp.js"
-import { getFirstCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getResumeByCandidateId } from "@/lib/db/resume/resume"
+import { getFirstCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getResumeByCandidateId } from "@/lib/db/resume/sql"
 
 // Server.
 const createServer = async () => {
