+++ src/app/api/resume/[candidateId]/education/route.ts
@@ -1,8 +1,8 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getEducationByCandidateId } from "@/lib/db/resume/education/education"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getEducationByCandidateId } from "@/lib/db/resume/education/sql"
 
 //
 // GET /api/resume/[candidateId]/education.
