+++ src/app/api/resume/[candidateId]/experience/[roleId]/[accomplishmentId]/route.ts
@@ -1,9 +1,9 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getAccomplishmentByCandidateIdRoleIdAndAccomplishmentId } from "@/lib/db/resume/experience/accomplishment"
-import { getRoleByCandidateIdAndRoleId } from "@/lib/db/resume/experience/role"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getAccomplishmentByCandidateIdRoleIdAndAccomplishmentId } from "@/lib/db/resume/experience/role/accomplishment/sql"
+import { getRoleByCandidateIdAndRoleId } from "@/lib/db/resume/experience/role/sql"
 
 //
 // GET /api/resume/[candidateId]/experience/[roleId]/[accomplishmentId].
