+++ src/app/api/resume/[candidateId]/experience/route.ts
@@ -1,8 +1,8 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getExperienceByCandidateId } from "@/lib/db/resume/experience/experience"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getExperienceByCandidateId } from "@/lib/db/resume/experience/sql"
 
 //
 // GET /api/resume/[candidateId]/experience.
