+++ src/app/api/resume/[candidateId]/skillSets/[skillSetId]/[skillId]/route.ts
@@ -1,9 +1,9 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getSkillByCandidateIdSkillSetIdAndSkillId } from "@/lib/db/resume/skillSets/skill"
-import { getSkillSetByCandidateIdAndSkillSetId } from "@/lib/db/resume/skillSets/skillSet"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getSkillByCandidateIdSkillSetIdAndSkillId } from "@/lib/db/resume/skillSets/skillSet/skill/sql"
+import { getSkillSetByCandidateIdAndSkillSetId } from "@/lib/db/resume/skillSets/skillSet/sql"
 
 //
 // GET /api/resume/[candidateId]/skillSets/[skillSetId]/[skillId].
