+++ src/app/api/resume/[candidateId]/skillSets/[skillSetId]/route.ts
@@ -1,8 +1,8 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getSkillSetByCandidateIdAndSkillSetId } from "@/lib/db/resume/skillSets/skillSet"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getSkillSetByCandidateIdAndSkillSetId } from "@/lib/db/resume/skillSets/skillSet/sql"
 
 //
 // GET /api/resume/[candidateId]/skillSets/[skillSetId].
