+++ src/app/api/resume/[candidateId]/skillSets/route.ts
@@ -1,8 +1,8 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
-import { getSkillSetsByCandidateId } from "@/lib/db/resume/skillSets/skillSets"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
+import { getSkillSetsByCandidateId } from "@/lib/db/resume/skillSets/sql"
 
 //
 // GET /api/resume/[candidateId]/skillSets/.
